M680X: Refactored cpu_type initialization for better readability.

+++ arch/M680X/M680XDisassemblerInternals.h
@@ -24,10 +24,10 @@ struct inst_pageX;
 typedef struct {
 	const struct inst_page1 *inst_page1_table;
 	const struct inst_pageX *inst_overlay_table[2];
-	unsigned int overlay_table_size[2];
+	size_t overlay_table_size[2];
 	uint8_t pageX_prefix[2];
 	const struct inst_pageX *inst_pageX_table[2];
-	unsigned int pageX_table_size[2];
+	size_t pageX_table_size[2];
 	const bool *reg_valid;
 	const bool *tfr_reg_valid;
 } cpu_tables;
